+++ src/Templating/test/GenerateTestProps.targets
@@ -1,12 +1,12 @@
 <Project>
   <Target Name="GenerateTestProps" BeforeTargets="CoreCompile">
+    <Message Text="$(MicrosoftAspNetCoreAppPackageVersion)" Importance="High" />
     <PropertyGroup>
       <PropsProperties>
         RestoreSources=$([MSBuild]::Escape($(RestoreSources)));
         RuntimeFrameworkVersion=$(RuntimeFrameworkVersion);
         MicrosoftNETSdkRazorPackageVersion=$(MicrosoftNETSdkRazorPackageVersion);
-        BundledAspNetCoreAppTargetFrameworkVersion=$(MicrosoftAspNetCoreAppPackageVersion.Split('.')[0]).$(MicrosoftAspNetCoreAppPackageVersion.Split('.')[1]);
-        BundledAspNetCoreAppPackageVersion=$(MicrosoftAspNetCoreAppPackageVersion)
+        MicrosoftAspNetCoreAppPackageVersion=$(MicrosoftAspNetCoreAppPackageVersion)
       </PropsProperties>
     </PropertyGroup>
